pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "weekly-product-review-pulse"
version = "0.1.0"
description = "Weekly Product Review Pulse agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-play-scraper>=1.2.7,<2",
  "fastapi>=0.111,<1",
  "google-api-python-client>=2.126,<3",
  "google-auth>=2.29,<3",
  "google-auth-oauthlib>=1.2,<2",
  "pydantic>=2.7,<3",
  "pydantic-settings>=2.2,<3",
  "PyYAML>=6.0.1,<7",
  "structlog>=24.1,<26",
  "typer>=0.12,<1",
  "uvicorn>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10,<2",
  "pytest>=8.2,<9",
  "ruff>=0.4,<1",
  "types-PyYAML>=6.0.12,<7",
]

[project.scripts]
pulse = "agent.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
exclude = ["agent/mcp/server.py"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
exclude = ["agent/mcp/server.py"]
python_version = "3.11"
packages = ["agent"]
strict = true

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["google_auth_oauthlib.*"]
ignore_missing_imports = true
